Simplify the pyspark template
