+++ project/pyspark/{{ cookiecutter.project_name }}/src/{{ cookiecutter.module_name }}/app.py
@@ -1,13 +1,17 @@
 import argparse
 
-from {{ cookiecutter.module_name }}.common.spark import ClosableSparkSession, SparkLogger
-from {{ cookiecutter.module_name }}.jobs import entrypoint
+from pyspark.sql import SparkSession
+from pyspark.sql import Row
+from pyspark.sql.dataframe import DataFrame
 
-# this import is required to discover the jobs
-# noinspection PyUnresolvedReferences
-from {{ cookiecutter.module_name }}.jobs import sample
+from {{ cookiecutter.module_name }}.common.spark import ClosableSparkSession, transform, SparkLogger
+from {{ cookiecutter.module_name }}.transformations.shared import add_ds
 
-if __name__ == "__main__":
+
+DataFrame.transform = transform
+
+
+def main():
     parser = argparse.ArgumentParser(description="{{ cookiecutter.project_name }}")
     parser.add_argument(
         "-d", "--date", dest="date", help="date in format YYYY-mm-dd", required=True
@@ -15,19 +19,70 @@ if __name__ == "__main__":
     parser.add_argument(
         "-e", "--env", dest="env", help="environment we are executing in", required=True
     )
-    parser.add_argument(
-        "-j",
-        "--jobs",
-        nargs="+",
-        dest="jobs",
-        help="jobs that need to be executed",
-        required=True,
-    )
     args = parser.parse_args()
 
     with ClosableSparkSession("{{ cookiecutter.project_name }}") as session:
-        logger = SparkLogger(session)
-        for job_name in args.jobs:
-            logger.info(f"Executing job {job_name}")
-            job = entrypoint.all[job_name]
-            job(session, args.env, args.date)
+        run(session, args.env, args.date)
+
+
+def run(spark: SparkSession, environment: str, date: str):
+    """Main ETL script definition.
+
+    :return: None
+    """
+    # execute ETL pipeline
+    logger = SparkLogger(spark)
+    logger.info(f"Executing job for {environment} on {date}")
+    data = extract_data(spark, date)
+    transformed = transform_data(data, date)
+    load_data(spark, transformed)
+
+
+def extract_data(spark: SparkSession, date: str) -> DataFrame:
+    """Load data from a source
+
+    :param spark: Spark session object.
+    :return: Spark DataFrame.
+    """
+    local_records = [
+        Row(user_id=1, event="login", browser="safari"),
+        Row(user_id=2, event="login", browser="chrome"),
+        Row(user_id=1, event="logout", browser="safari"),
+        Row(user_id=1, event="logout", browser="safari"),
+        Row(user_id=2, event="expired", browser="IE6"),
+        Row(user_id=3, event="login", browser="firefox"),
+    ]
+    return spark.createDataFrame(local_records)
+
+
+def transform_data(data: DataFrame, date: str) -> DataFrame:
+    """Transform original dataset.
+
+    :param data: Input DataFrame.
+    :param date: The context date
+    :return: Transformed DataFrame.
+    """
+    return data.transform(add_ds(date)).dropDuplicates()
+
+
+def load_data(spark: SparkSession, data: DataFrame):
+    """Writes the output dataset to some destination
+
+    :param data: DataFrame to write.
+    :return: None
+    """
+    data.show()
+
+    # Uncomment the following block to write to a compatible catalog
+    # spark.catalog.setCurrentDatabase("DEFAULT_DB")
+    # (
+    #     data.coalesce(1)
+    #     .write.partitionBy("df")
+    #     .mode("overwrite")
+    #     .format("parquet")
+    #     .saveAsTable("sample")
+    # )
+
+
+if __name__ == "__main__":
+    main()
